V1.2
All functionality for the JDBC driver have been added other than this
specification:

-Enter a valid SELECT/UPDATE/INSERT/DELETE SQL command. The system
should take the corresponding action, and return and display the valid
results. For a SELECT query, display the answers. For the other types
of queries, give enough information about the status of the execution
of the query. For instance, for an UPDATE query, show the user how many
records have been successfully changed.

It should be easy to implement if anyone else wishes to program it.